% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.functions.R
\name{addReArrangmentMatrix}
\alias{addReArrangmentMatrix}
\title{addReArrangmentMatrix}
\usage{
addReArrangmentMatrix(
  object,
  vdj.df,
  target.col = "reAr",
  barcodes = "BC",
  mtr.name = NULL,
  set.active = T
)
}
\arguments{
\item{vdj.df}{The data frame with barcodes and the ReArrangment parameters}

\item{target.col}{Colname of the target enrichment colum}

\item{barcodes}{Colname of the barcodes}

\item{mtr.name}{The name of the importet dataset}
}
\value{
SPATA object
}
\description{
This function transformes a input to a expression matrix and import it to the spata object
}
\examples{

}
\author{
Dieter Henrik Heiland
}
