% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.import.R
\name{create.IAF}
\alias{create.IAF}
\title{create.IAF}
\usage{
create.IAF(Binner.file, sam.file)
}
\value{

}
\description{
This function creates a TCR aligned file from the alignment (spatial bc) and the TCR-sequences
}
\examples{

}
\author{
Dieter Henrik Heiland
}
