% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.get.R
\name{GetClonalityAnalysis}
\alias{GetClonalityAnalysis}
\title{GetClonalityAnalysis}
\usage{
GetClonalityAnalysis(object, constant = "TRB")
}
\value{
SPATA object
}
\description{
This function import the VDJ data from the SPTCR-Pipline or IgBlast
}
\author{
Dieter Henrik Heiland
}
