% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.get.R
\name{getSpatialRegression}
\alias{getSpatialRegression}
\title{getSpatialRegression}
\usage{
getSpatialRegression(object, features, model, smooth = F, normalize = F)
}
\arguments{
\item{object}{SPATA object.}

\item{features}{features to perform spatial regression.}

\item{model}{The model for spatial weighted regression. classical:Pearson (Not recommended), CCA model, Spatial Lag model: lmSLX, dist}

\item{smooth}{Kernel.}

\item{normalize}{Normalize Input (Not recommended).}
}
\value{
Correlation matrix of NxN input
}
\description{
Wrapper function for spatial weighted regression analysis
}
\author{
Dieter Henrik Heiland
}
