% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.get.R
\name{GetSPTCR}
\alias{GetSPTCR}
\title{GetSPTCR}
\usage{
GetSPTCR(object, constant = NULL)
}
\value{
SPATA object
}
\description{
This function import the VDJ data from the SPTCR-Pipline
}
\author{
Dieter Henrik Heiland
}
