% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.get.R
\name{GetVDJ}
\alias{GetVDJ}
\title{GetVDJ}
\usage{
GetVDJ(object, constant = NULL)
}
\value{
SPATA object
}
\description{
This function import the VDJ data from the SPTCR-Pipline or IgBlast
}
\author{
Dieter Henrik Heiland
}
