% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.hlp.R
\name{hlpFlattenDF}
\alias{hlpFlattenDF}
\title{hlpFlattenDF}
\usage{
hlpFlattenDF(data.frame)
}
\value{
SPATA object
}
\description{
This function import the VDJ data from the SPTCR-Pipline or IgBlast
}
\author{
Dieter Henrik Heiland
}
