% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.hlp.R
\name{hlpList2Array}
\alias{hlpList2Array}
\title{hlpList2Array}
\usage{
hlpList2Array(input.list)
}
\value{
SPATA object
}
\description{
This function import the VDJ data from the SPTCR-Pipline or IgBlast
}
\author{
Dieter Henrik Heiland
}
