% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.import.R
\name{import.sam.files}
\alias{import.sam.files}
\title{import.sam.files}
\usage{
import.sam.files(object, IAF.file)
}
\value{

}
\description{
Add on SPATAImmune
}
\examples{

}
\author{
Dieter Henrik Heiland
}
