% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.functions.R
\name{ImportSPTCRseq}
\alias{ImportSPTCRseq}
\title{ImportSPTCRseq}
\usage{
ImportSPTCRseq(object, sample_name, path)
}
\arguments{
\item{object}{The data frame with barcodes and the ReArrangment parameters}

\item{path}{The path to the SPTCR-seq pipeline output}
}
\value{
SPATA object
}
\description{
This function import the VDJ data from the SPTCR-Pipline or IgBlast
}
\examples{

}
\author{
Dieter Henrik Heiland
}
