% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.functions.R
\name{ImportVDJ}
\alias{ImportVDJ}
\title{ImportVDJ}
\usage{
ImportVDJ(
  object,
  barcodes.df,
  IgBlast.df,
  Pychoper = F,
  Filter.cdr3 = T,
  ram = 8,
  cors = 16
)
}
\arguments{
\item{object}{The data frame with barcodes and the ReArrangment parameters}

\item{barcodes.df}{Data frame of the barcodes, with the following cols: c("ID", "barcodes"): ID: Sequencing ID, barcodes: The annotated Barcode}

\item{IgBlast.df}{Optimized IgBlast output with the following cols: c("ID", "c", "v", "d", "j", "CDR3"). ID: Sequencing ID, c: Constant Region, v,d,j calls, and the CDR3 sequence}

\item{Pychoper}{If the Pychopper option was used before IgBlast (in the pipline)}
}
\value{
SPATA object
}
\description{
This function import the VDJ data from the SPTCR-Pipline or IgBlast
}
\examples{

}
\author{
Dieter Henrik Heiland
}
