% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.run.R
\name{modelAA2vec}
\alias{modelAA2vec}
\title{modelAA2vec}
\usage{
modelAA2vec()
}
\value{
SPATA object
}
\description{
This function import the VDJ data from the SPTCR-Pipline or IgBlast
}
\author{
Dieter Henrik Heiland
}
