% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.plots.R
\name{plot3DClonality2ST}
\alias{plot3DClonality2ST}
\title{plot3DClonality2ST}
\usage{
plot3DClonality2ST(
  object,
  color_by,
  pt.bot = 3,
  pt.up = 1,
  random.spots = 10,
  line.size = 0.3,
  select.cluster = NULL,
  constant = "TRB",
  angle = list(x = 1.87, y = 0.88, z = -0.64)
)
}
\value{

}
\description{
SPATA-Immuno
}
\examples{

}
\author{
Dieter Henrik Heiland
}
