% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.plots.R
\name{plotAAlogo}
\alias{plotAAlogo}
\title{plotAAlogo}
\usage{
plotAAlogo(object, motif.length = 8, cluster.select = 2, constant = "TRB")
}
\value{

}
\description{
SPATA-Immuno
}
\examples{

}
\author{
Dieter Henrik Heiland
}
