% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.plots.R
\name{plotAggregatedDimred}
\alias{plotAggregatedDimred}
\title{plotAggregatedDimred}
\usage{
plotAggregatedDimred(
  object,
  factor = 20,
  pt.alpha = 0.8,
  pt.size = 0.5,
  return.coords = F,
  constant
)
}
\value{

}
\description{
SPATA-Immuno
}
\examples{

}
\author{
Dieter Henrik Heiland
}
