% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.plots.R
\name{plotBarCompare}
\alias{plotBarCompare}
\title{plotBarCompare}
\usage{
plotBarCompare(plot_df, Feat, compare_to, order = NULL)
}
\value{

}
\description{
SPATA-Immuno
}
\examples{

}
\author{
Dieter Henrik Heiland
}
