% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.plots.R
\name{plotRiver}
\alias{plotRiver}
\title{River Plot}
\usage{
plotRiver(data, var1, var2, color1 = NULL, color2 = NULL)
}
\value{

}
\description{
SPATA-Immuno
}
\examples{

}
\author{
Dieter Henrik Heiland
}
