% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.plots.R
\name{plotVDJCirclize}
\alias{plotVDJCirclize}
\title{plotVDJCirclize}
\usage{
plotVDJCirclize(
  object,
  select.c = "TRB",
  arrange.by = NULL,
  top = 20,
  output = getwd()
)
}
\value{

}
\description{
SPATA-Immuno
}
\examples{

}
\author{
Dieter Henrik Heiland
}
