% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.run.R
\name{runClonalityAnalysis}
\alias{runClonalityAnalysis}
\title{runClonalityAnalysis}
\usage{
runClonalityAnalysis(object, motif.length = 8, constant = "TRB")
}
\value{
SPATA object
}
\description{
This function import the VDJ data from the SPTCR-Pipline or IgBlast
}
\author{
Dieter Henrik Heiland
}
