% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.run.R
\name{runKmers}
\alias{runKmers}
\title{runKmers}
\usage{
runKmers(seq, kmer.length = 4, kmer.nr = NULL, overlap = 1)
}
\value{
SPATA object
}
\description{
This function import the VDJ data from the SPTCR-Pipline or IgBlast
}
\author{
Dieter Henrik Heiland
}
