% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.run.R
\name{runSpaceXR}
\alias{runSpaceXR}
\title{runSpaceXR}
\usage{
runSpaceXR(Seurat, cluster, SPATA)
}
\arguments{
\item{Seurat}{Seurat data (scRNA-seq).}

\item{cluster}{Meta data parameter.}

\item{object}{SPATA object.}
}
\value{
SPATA object with integrated clusters (from Seurat)
}
\description{
Single Cell integration
}
\author{
Dieter Henrik Heiland
}
